normalized authority weight and pagerank
Q1
Consider the HITS algorithm. Assume that there are four web sites A, B, C, and D. The interconnection among them are shown as follows:

Screen Shot 2022-03-27 at 9.19.40 PM.png

What are the normalized authority weight and the normalized hub weight for each site in the first three iterations?
Let the final ranking score of a site be the sum of its authority weight and its hub weight (using the weight vector computed after three iterations in the above question. Please rank the sites in descending order of the ranking scores. (If there are ties, please rank them alphabetically)
Q2
Consider the simple PageRank definition. Let  denote the set of inlink neighbors of , denote the outdegree of , then the PageRank of  is . Following this definition, please show the computed PageRank score of each node in the following network for the first two iterations. Note that  is initialized as 1.

Screen Shot 2022-03-27 at 9.24.21 PM.png