bagging and boosting
Q1
Please briefly compare bagging and boosting from the following perspectives: (a) assumption; (b) construction process; (c) final aggregation of classifiers.

Q2
Assume that at threshold , a binary classification model  classifies 60 out of 80 test cases as positive; while 12 out of 60 predicted positives are actually negative.  has missed 5 cases that are actually positive.

(a) Please compute the precision and recall of .
(b) what are the coordinates of  on a ROC curve?