svm
Q1
Consider the following data points labeled in , where is the label. For example, (3,-1) is labeled as positive, while (1,0) is labeled as negative. We would like to discover an SVM that accurately discriminates the two classes.

ID
1	3	1	1
2	3	-1	1
3	6	1	1
4	6	-1	1
5	1	0	-1
6	0	1	-1
7	0	-1	-1
8	-1	0	-1
(a) Is a linear SVM sufficient for this problem? In other words, are the data points linearly separable?

(b) Please list all the support vectors.

(c) Find the decision boundary.

(d) Please use the SVM computed from above to classify the following data points: (0,0) and (3,3).

Q2
We are given two data records with 2 different timestamps, where  and  are input variables,  is the output variable.

Timestamps
0.3	0.6	0.2
0.1	1.0	0.4
Consider the traditional LSTM model. Initially, we have the following internal weight vectors and bias variables as follows:

In the model, for each timestamp , we have the following state variables: (1) forget gate variable ; (2) input gate variable ; (3) input activation variable ; (4) internal state variable ; (5) output gate variable ; and (6) final output variable . Consider the input forward propagation step only.

(a) Let , what are the values of the above state variables at  and ? Please show each answer up to 4 decimal places.

(b) Based on the computation in Q1, what are the errors of the final output variables at  and ? Please show each answer up to 4 decimal places.