Consider the following four  data points: (2, 2), (4, 4), (1, 5) and (5, 1). We can make use of the KL-Transform to find a transformed subspace containing a cluster. Let  be the total number of dimensions in the original space and  be the total number of dimensions in the projected subspace. Please illustrate the KL-transform technique with the above example when  and .  Note: You need to show the covariance matrix, eigenvalues and eigenvectors, and the transformation step in your solution.     Q2 Consider the following 6 2-dimensional points:   Assume that we need to find the outlier in this tiny data set. Please apply the Manhattan distance to answer the following questions. (Manhattan distance is defined as follows: in a plane with p1 at (x1, y1) and p2 at (x2, y2), it is |x1 - x2| + |y1 - y2|)  Consider the distance-based model introduced in class, let  be 4, and the size threshold be 3, which data point(s) would be identified as outlier(s)? Consider the density-based model introduced in class, let , what is the local outlier factor of data point  and , respectively?
Consider the following four 2-dimensional data points: (2, 2), (4, 4), (1, 5) and (5, 1). We can make use of the KL-Transform to find a transformed subspace containing a cluster. Let  be the total number of dimensions in the original space and  be the total number of dimensions in the projected subspace. Please illustrate the KL-transform technique with the above example when  and .

Note: You need to show the covariance matrix, eigenvalues and eigenvectors, and the transformation step in your solution.

Q2
Consider the following 6 2-dimensional points:

Assume that we need to find the outlier in this tiny data set. Please apply the Manhattan distance to answer the following questions. (Manhattan distance is defined as follows: in a plane with p1 at (x1, y1) and p2 at (x2, y2), it is |x1 - x2| + |y1 - y2|)

Consider the distance-based model introduced in class, let  be 4, and the size threshold be 3, which data point(s) would be identified as outlier(s)?
Consider the density-based model introduced in class, let , what is the local outlier factor of data point  and , respectively?