k means algorithm
Q1
Consider the following eight two-dimensional data points:

You are required to use the -means algorithm to cluster these points. You need to show the information about each final cluster (including the mean of the cluster and all data points in this cluster).

If  and the initial means are (10,1) and (10,30), what is the output of the algorithm?
If  and the initial means are (10,1), (10,30), and (3,10), what is the output of the algorithm?
Q2
What are the advantages and disadvantages of the -means algorithm? For each disadvantage, please also give a suggestion to enhance the algorithm.

Q3
The following matrix shows the pairwise distances between any two points among eight points.

Q3

Please use the agglomerative approach to group these points with distance group average linkage. Draw the corresponding dendrogram for the clustering
Suppose that we want to find 5 clusters. According to the dendrogram derived in the above question, please state the 5 clusters. For each cluster, please include all data points involved.