Fp Tree growth
This assignment dues on 1/31/2022, by 23:59. The late submission policy can be found in the Course Syllabus.
Please make sure your upload is clear and readable if it is handwritten.
We are given four items, namely , , , and . Their corresponding unit profits are , , , and . The following shows five transactions with these items. Each row corresponds to a transaction where a non-negative integer shown in the row corresponds to the total number of occurrences of the correspondence item present in the transaction.

Transaction Table
0	0	3	2
3	4	0	0
0	0	1	3
1	0	3	5
6	0	0	0
The frequency of an itemset in a row is defined to be the minimum of the number of occurrences of all items in the itemset. For example, itemset  in the first row has frequency equals 2. But, itemset  in the third row has a frequency as 1.
The frequency of an itemset in the dataset is defined to be the sum of the frequencies of the itemset in all rows in the dataset. For example, itemset  has frequency 2+0+1+3+0=6.
Let  be a function defined on an itemset ( will be specified later). One example of this functions is . In this example, if , then .

The profit of an itemset in the dataset is defined to be the product of the frequency of this itemset in the dataset and . For example, itemset has profit .
Answer the following two questions based on the above problem setting.

Q1
Assume , where  is the size of . Suppose that we know . We want to find all itemsets with profit of at least 50. Can the Apriori Algorithm be adapted to find these itemsets? If yes, please write down the pseudo-code* and illustrate it with the above example. If no, please explain why. In this case, please also design an algorithm for this problem and write down the pseudo-code.

Q2
Assume . Suppose that we know . We want to find all itemsets with profit of at least 50. Can the Apriori Algorithm be adapted to find these itemsets? If yes, please write down the pseudo-code and illustrate it with the above example. If no, please explain why. In this case, please also design an algorithm for this problem and write down the pseudo-code.

Note
*pseudo-code is an informal high-level description of the algorithm. If you are not familiar with this type of algorithm abstraction, you can describe the process in plain language. Please DO NOT submit the real code!

Q3
Consider the following transactions. Each row corresponds to a transaction where 1 corresponds to the presence of an item and 0 corresponds to an absence.

1	0	0	1
0	1	0	0
1	0	1	1
1	0	1	1
Let the minimum support threshold be 2. Please illustrate the four steps of the FP-tree growth algorithm with the above example.

Q4 (COMP7118 Only)
Consider the FP-tree growth algorithm. If we shuffle the order of transactions in the second step, will the final tree structure be different? Please explain your answer.

Q5 (COMP8118 Only)
Please explain how the FP-tree growth algorithm can be adopted to find closed frequent itemsets and maximal frequent itemsets.